Refactor: extract scalaz runtime traits
